App: extract output settings into module
